# docs/swagger.yaml
basePath: /
definitions:
  fiber.Map:
    additionalProperties: true
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  handlers.SignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.About:
    properties:
      address:
        type: string
      age:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      surname:
        type: string
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  models.Equipment:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      purpose:
        type: string
      updatedAt:
        type: string
    type: object
  models.Group:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/models.Users'
        type: array
    type: object
  models.Gym:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      equipment:
        items:
          $ref: '#/definitions/models.Equipment'
        type: array
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.Sessions:
    properties:
      createdAt:
        type: string
      date:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      durationInMinutes:
        type: string
      exerciseType:
        type: string
      id:
        type: integer
      isWorkoutOfDay:
        type: boolean
      localisation:
        type: string
      movesNumber:
        type: string
      rating:
        type: string
      reviews:
        type: string
      setsNumber:
        type: string
      timeLeftInHour:
        type: string
      updatedAt:
        type: string
      workOutTitle:
        type: string
    type: object
  models.Users:
    properties:
      about:
        $ref: '#/definitions/models.About'
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      groups:
        description: RoleID   uint    `json:role_id gorm:"default:1"`
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: integer
      password:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: Gym API for users.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gym API
  version: "1.0"
paths:
  /abouts:
    post:
      consumes:
      - application/json
      description: Create a new about
      parameters:
      - description: About
        in: body
        name: about
        required: true
        schema:
          $ref: '#/definitions/models.About'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.About'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Create a new about
      tags:
      - abouts
  /abouts/:
    get:
      consumes:
      - application/json
      description: Get all the Abouts from the Database
      parameters:
      - description: About
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.About'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Show all the Abouts
      tags:
      - abouts
  /abouts/{id}:
    delete:
      description: Delete an about by ID
      parameters:
      - description: About ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete an about by ID
      tags:
      - abouts
    get:
      description: Get an about by ID
      parameters:
      - description: About ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.About'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Retrieve an about by ID
      tags:
      - About
    put:
      consumes:
      - application/json
      description: Update an about by ID
      parameters:
      - description: About ID
        in: path
        name: id
        required: true
        type: integer
      - description: About
        in: body
        name: about
        required: true
        schema:
          $ref: '#/definitions/models.About'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.About'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update an about by ID
      tags:
      - abouts
  /gyms:
    get:
      description: Get a list of all gyms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Gym'
            type: array
      summary: Get all gyms
      tags:
      - Gyms
    post:
      consumes:
      - application/json
      description: Create a new gym with name and location
      parameters:
      - description: Gym
        in: body
        name: gym
        required: true
        schema:
          $ref: '#/definitions/models.Gym'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Gym'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Create a new gym
      tags:
      - Gyms
  /gyms/{id}:
    delete:
      description: Delete a gym by its ID
      parameters:
      - description: Gym ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete a gym by ID
      tags:
      - Gyms
    get:
      description: Get a gym by its ID
      parameters:
      - description: Gym ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Gym'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get a gym by ID
      tags:
      - Gyms
    put:
      consumes:
      - application/json
      description: Update a gym's details by its ID
      parameters:
      - description: Gym ID
        in: path
        name: id
        required: true
        type: integer
      - description: Gym
        in: body
        name: gym
        required: true
        schema:
          $ref: '#/definitions/models.Gym'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Gym'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Update a gym by ID
      tags:
      - Gyms
  /logout:
    post:
      description: Invalidate the user's session or token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Log out a user
      tags:
      - Signup & Login
  /sessions:
    get:
      description: Retrieve all sessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Sessions'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all sessions
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      description: Create a new session
      parameters:
      - description: Session
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/models.Sessions'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Sessions'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new session
      tags:
      - Sessions
  /sessions/{id}:
    delete:
      description: Delete a session by ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete a session by ID
      tags:
      - Sessions
    get:
      description: Retrieve a session by ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sessions'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get a session by ID
      tags:
      - Sessions
    put:
      consumes:
      - application/json
      description: Update a session's details by ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      - description: Session
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/models.Sessions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sessions'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Update a session by ID
      tags:
      - Sessions
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user with email and password
      parameters:
      - description: Sign up request
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/handlers.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Sign up a new user
      tags:
      - Signup & Login
  /users:
    get:
      description: Retrieve all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Users'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Users'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Users'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete a user by ID
      tags:
      - Users
    get:
      description: Retrieve a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Users'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get a user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update a user's details by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Users'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Update a user by ID
      tags:
      - Users
swagger: "2.0"